Splitting SelectExpand E2E tests into two separate classes
While trying to add complex types to the test model for SelectExpandTests, I
ran into an issue with EF where some existing tests failed due to EF's own EDM
model validation. This was happening because it assumes the new complex type
is actually an entity type, even though it doesn't have a key.

My solution is just to factor out the EF-related tests into their own class,
SelectExpandEFTests. This should allow the addition of complex types to the
SelectExpandTests test model without breaking any of the existing tests.
