+++ test/E2ETest/Microsoft.Test.E2E.AspNet.OData/Build.AspNetCore/Microsoft.Test.E2E.AspNetCore.OData.csproj
@@ -1403,6 +1403,9 @@
     <Compile Include="..\QueryComposition\SelectExpandTests.cs">
       <Link>QueryComposition\SelectExpandTests.cs</Link>
     </Compile>
+    <Compile Include="..\QueryComposition\SelectExpandEFTests.cs">
+      <Link>QueryComposition\SelectExpandEFTests.cs</Link>
+    </Compile>
     <Compile Include="..\QueryComposition\SelectTypeExplosionTests.cs">
       <Link>QueryComposition\SelectTypeExplosionTests.cs</Link>
     </Compile>
